Remove unnecessary data and cleaned everything up into one file.
